requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
cachetools==5.5.2
django==4.2.22
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.40.3
importlib-metadata==8.5.0
Markdown==3.7
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
PyJWT==2.9.0
python-decouple==3.8
redis==6.1.1
rsa==4.9.1
sqlparse==0.5.3
typing-extensions==4.13.2
zipp==3.20.2
